Add files via upload
This is a smile detector final project of the Big Data analytics course at Cisco Neta Academy. It uses Raspberry Pi with its Pi Cam that runs Python 3. I used the OpenCV module for image processing and taking snapshots when detecting smiles. The model is trained with straight and smiling photos cropped to the face and has a higher accuracy rate.